Implement friends free-slots feature and refactor UI
Add domain models and use-cases for friends' free slots (FreeSlotSelection, FriendsFreeSlotsRoomRecommendations, FindRoomsForFreeSlotsUseCase, LoadFriendsFreeSlotsForWeekUseCase) and wire room search integration. Introduce Riverpod notifiers and state for friends free slots, move controller files to notifiers, and add presentation widgets (content, controls, grid, theme) to split the large page. Refactor FriendsFreeSlotsPage to a ConsumerWidget that uses the new notifier, update imports accordingly, and navigate to RecommendedRoomsPage using merged room results and time filter options. Also include small schedule repository changes and provider updates to support the new flows. This refactor extracts UI logic into reusable components and centralizes async/business logic in use-cases and notifiers.
