+++ lib/features/friendships/domain/entities/friends_free_slots_room_recommendations.dart
@@ -1,0 +1,13 @@
+import 'package:andespace/features/rooms/domain/entities/room_search.dart';
+
+class FriendsFreeSlotsRoomRecommendations {
+  const FriendsFreeSlotsRoomRecommendations({
+    required this.rooms,
+    required this.timeFilterOptions,
+    required this.description,
+  });
+
+  final List<RoomSearchItem> rooms;
+  final List<MatchingWindow> timeFilterOptions;
+  final String description;
+}
