+++ lib/features/friendships/presentation/pages/add_friends_page.dart
@@ -2,8 +2,8 @@ import 'package:andespace/core/navigation/app_routes.dart';
 import 'package:andespace/core/navigation/app_tab.dart';
 import 'package:andespace/features/auth/presentation/notifiers/auth_notifier.dart';
 import 'package:andespace/features/friendships/domain/entities/friendship_request.dart';
-import 'package:andespace/features/friendships/presentation/controllers/friendships_notifier.dart';
-import 'package:andespace/features/friendships/presentation/controllers/friendships_state.dart';
+import 'package:andespace/features/friendships/presentation/notifiers/friendships_notifier.dart';
+import 'package:andespace/features/friendships/presentation/notifiers/friendships_state.dart';
 import 'package:andespace/features/friendships/presentation/providers/friendships_providers.dart';
 import 'package:andespace/shared/theme/app_theme_extension.dart';
 import 'package:andespace/shared/widgets/app_scaffold.dart';
