+++ lib/features/friendships/presentation/pages/friends_page.dart
@@ -3,8 +3,8 @@ import 'package:andespace/core/navigation/app_tab.dart';
 import 'package:andespace/features/auth/domain/entities/user_status.dart';
 import 'package:andespace/features/auth/presentation/notifiers/auth_notifier.dart';
 import 'package:andespace/features/friendships/domain/entities/friend.dart';
-import 'package:andespace/features/friendships/presentation/controllers/friendships_notifier.dart';
-import 'package:andespace/features/friendships/presentation/controllers/friendships_state.dart';
+import 'package:andespace/features/friendships/presentation/notifiers/friendships_notifier.dart';
+import 'package:andespace/features/friendships/presentation/notifiers/friendships_state.dart';
 import 'package:andespace/features/friendships/presentation/pages/add_friends_page.dart';
 import 'package:andespace/features/friendships/presentation/pages/friend_schedule_page.dart';
 import 'package:andespace/features/friendships/presentation/pages/friends_free_slots_page.dart';
